requirements.txt
python-dotenv
pyniryo
scapy
